# Remove unneded code in `ConstantArrayType::hasOffsetValueType()`