was added in 28c5ff61, but everything seems to be still fine without it. not sure yet if this changes something performance-wise, I hope CI tells me